agents/host.py
from typing import Dict, List, Any
from langchain.agents import AgentExecutor
from langchain.schema import AgentAction, AgentFinish

class HostAgent:
    """Agent chịu trách nhiệm cho Auto-mode và Q&A chung"""
    
    def __init__(self):
        self.name = "Host"
        self.description = "Agent chịu trách nhiệm điều phối và trả lời các câu hỏi chung"
    
    async def run(self, input_data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Xử lý input và trả về kết quả
        
        Args:
            input_data: Dictionary chứa thông tin input
            
        Returns:
            Dictionary chứa kết quả xử lý
        """
        # TODO: Implement logic xử lý
        return {"status": "success", "message": "Host agent processed the request"} 